just save

+++ src/git/doomshade/professions/data/cache/CacheUtils.java
@@ -6,55 +6,78 @@ import java.io.*;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.Collections;
-import java.util.HashMap;
 
 public class CacheUtils {
-
-    private static final HashMap<String, ObjectOutputStream> CACHE = new HashMap<>();
 
     private CacheUtils() {
     }
 
+    /**
+     * Caches single data to a file with given folder.
+     *
+     * @param data     the data to cache
+     * @param fileName the file name
+     * @param folder   the folder
+     * @param <T>      the data type
+     * @throws IOException if cache was unsuccessful
+     */
     public static <T extends Cacheable> void cache(T data, String fileName, String folder) throws IOException {
         cache(Collections.singleton(data), fileName, folder);
     }
 
+    /**
+     * Caches data to a file with given folder.
+     *
+     * @param data     the data to cache
+     * @param fileName the file name
+     * @param folder   the folder
+     * @param <T>      the data type
+     * @throws IOException if cache was unsuccessful
+     */
     public static <T extends Cacheable> void cache(Collection<T> data, String fileName, String folder) throws IOException {
         File dir = new File(Professions.getInstance().getCacheFolder() + File.separator + folder);
 
-        boolean append = true;
+        // create folder and file if they don't exist
         if (!dir.isDirectory()) {
             dir.mkdirs();
-            append = false;
         }
 
         File file = new File(dir, fileName);
         if (!file.exists()) {
             file.createNewFile();
-            append = false;
         }
 
-        if (append)
-            appendCache(data, file);
-        else
-            setupCache(data, file);
+        setupCache(data, file);
     }
 
-    private static <T extends Cacheable> void appendCache(Collection<T> data, File file) throws IOException {
-        ObjectOutputStream oos = CACHE.get(file.getAbsolutePath());
-        for (T d : data)
-            oos.writeObject(d);
-        oos.flush();
-    }
-
+    /**
+     * Sets up cache file, writing the header first and then the data
+     *
+     * @param data the data to cache
+     * @param file the file
+     * @param <T>  the data type
+     * @throws IOException if setting up was unsuccessful
+     */
     private static <T extends Cacheable> void setupCache(Collection<T> data, File file) throws IOException {
-        ObjectOutputStream oos = new ObjectOutputStream(new FileOutputStream(file, true));
-        for (T d : data)
-            oos.writeObject(d);
-        oos.flush();
-        CACHE.put(file.getAbsolutePath(), oos);
+        try (ObjectOutputStream oos = new ObjectOutputStream(new FileOutputStream(file, true))) {
+            // save time, when this cache was created
+            oos.writeObject(new CacheHeader());
+
+            // then write data
+            for (T d : data)
+                oos.writeObject(d);
+        }
     }
 
+    /**
+     * Reads the cache. The first index of this collection is
+     *
+     * @param fileName the file name
+     * @param folder   the file folder
+     * @param <T>      the data type
+     * @return the collection of deserialized cacheable data
+     * @throws IOException if deserialization was unsuccessful
+     */
     @SuppressWarnings("all")
     public static <T extends Cacheable> Collection<T> readCache(String fileName, String folder) throws IOException {
         Collection<T> coll = new ArrayList<>();
@@ -71,48 +94,34 @@ public class CacheUtils {
         return coll;
     }
 
+    /**
+     * Clears the cache folder
+     *
+     * @param folder the folder to clear
+     */
     public static void clearCache(String folder) {
+
+        // ignore empty folder, this would delete the whole cache folder
+        if (folder.isEmpty()) return;
+
+
         File files = new File(Professions.getInstance().getCacheFolder(), folder);
 
         if (files.isDirectory()) {
             for (File f : files.listFiles()) {
-                ObjectOutputStream oos = CACHE.get(f.getAbsolutePath());
-                if (oos != null) {
-                    try {
-                        oos.close();
-                    } catch (IOException e) {
-                        Professions.logError(e);
-                    }
-                }
                 f.delete();
             }
         }
     }
 
     public static void clearCache(String fileName, String folder) {
+
+        // ignore empty folder/name
+        if (folder.isEmpty() || fileName.isEmpty()) return;
         File f = new File(Professions.getInstance().getCacheFolder() + File.separator + folder, fileName);
 
         if (f.exists()) {
-            ObjectOutputStream oos = CACHE.get(f.getAbsolutePath());
-            if (oos != null) {
-                try {
-                    oos.close();
-                } catch (IOException e) {
-                    Professions.logError(e);
-                }
-            }
             f.delete();
         }
-    }
-
-    public static void closeCacheFiles() {
-        CACHE.forEach((x, y) -> {
-            try {
-                y.flush();
-                y.close();
-            } catch (IOException e) {
-                e.printStackTrace();
-            }
-        });
     }
 }
